% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadDecadalForecast.R
\name{adjustDates.forecast}
\alias{adjustDates.forecast}
\title{Adjust dates in forecast data}
\usage{
adjustDates.forecast(foreTimePars)
}
\arguments{
\item{foreTimePars}{A list of elements as returned by \code{\link{getRunTimeDomain}}}
}
\description{
Adjust dates in forecast data
}
\author{
S. Herrera
}

