% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{C4R.datasets}
\alias{C4R.datasets}
\title{Show standard public datasets}
\usage{
C4R.datasets(pattern = "")
}
\arguments{
\item{pattern}{Optional. Pattern in the dataset name as passed to function \code{\link{grep}} (case-insensitive).}
}
\value{
The datasets table, in the form of a \code{data.frame}
}
\description{
Show standard public datasets
}
\note{
The function assumes that the user has read permission to the package installation directory
}
\examples{
# Default built-in datasets
str(C4R.datasets())
}
\author{
M Iturbide
}
