% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{C4R.vocabulary}
\alias{C4R.vocabulary}
\title{Show UDG vocabulary}
\usage{
C4R.vocabulary()
}
\value{
The vocabulary table, in the form of a data.frame
}
\description{
Access the installed user's vocabulary
}
\note{
The function assumes that the user has read permission to the package installation directory
}
\examples{
# Default built-in vocabulary
(voc <- C4R.vocabulary())
voc[grep("^ta", voc$identifier), ]
}
\references{
Standard name table of the CF convention: http://cfconventions.org/standard-names.html
}
\seealso{
C4R.vocabulary.update, for the inclusion of new standard variables defined by the user
}
\author{
J Bedia
}
