% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaryLookup.R
\name{check.dictionary}
\alias{check.dictionary}
\title{Check for dictionary options}
\usage{
check.dictionary(dataset, var, dictionary, time)
}
\arguments{
\item{dataset}{dataset}

\item{var}{var}

\item{dictionary}{dictionary}

\item{time}{time}
}
\value{
a list with actual shortName and dic definition
}
\description{
Check dictionary argument and transform variable name and generate dic details accordingly
}
\author{
J. Bedia
}
\keyword{internal}

