% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{codigo.monokai}
\alias{codigo.monokai}
\title{Returns HTML for code}
\usage{
codigo.monokai(id, height)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{height}{The height of the input, e.g. '400px', or '100vh'.}
}
\value{
HTML
}
\description{
Returns HTML for code
}
\examples{
codigo.monokai("id", "70vh")

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
