% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dictionaryTransformForecast.R
\name{dictionaryTransformForecast}
\alias{dictionaryTransformForecast}
\title{Performs variable transformation}
\usage{
dictionaryTransformForecast(dic, foreTimePars, mdArray)
}
\arguments{
\item{dic}{Dictionary line for the variable, as returned by \code{dictionaryLookup.ECOMS}}

\item{foreTimePars}{A list of time selection parameters, as returned by \code{getTimeDomain}}

\item{mdArray}{A n-dimensional array, as returned by \code{makeSubset}}
}
\value{
The transformed n dimensional array of data. See details.
}
\description{
Performs variable transformation according to dictionary specifications.
Sub-routine of the \code{loadSeasonalForecast} methods.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}

