% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getForecastTimeDomain.R
\name{getForecastTimeDomain}
\alias{getForecastTimeDomain}
\title{Selection of time slices of forecast datasets}
\usage{
getForecastTimeDomain(grid, dataset, dic, runTimePars, time, aggr.d, aggr.m)
}
\arguments{
\item{grid}{a java \sQuote{GeoGrid}}

\item{dataset}{character string of the dataset}

\item{dic}{dictionary information}

\item{runTimePars}{A list of elements as returned by \code{getRunTimeDomain*}}

\item{time}{Verification time defined as a character string (e.g. \dQuote{06} for data
verifying at 06:00:00). Only applies for sub-daily datasets.}

\item{aggr.d}{Aggregation function for subdaily to daily}

\item{aggr.m}{Aggregation function from daily to monthly}
}
\value{
A list with the following elements:
\item{forecastDates}{A list with POSIXlt dates defining the start and end of the 
representative verification time. If start and end are identical, the variable is instantaneous
and therefore the representative time interval is 0}
\item{foreTimeRangesList}{A list of length \emph{i} containing the java ranges defining the
forecast times selected along the \emph{i-th} run time axis.}
\item{foreTimeShift}{Integer value (java format) giving the shift to start reading in the time axis}
\item{foreTimeStride}{Integer value (java format) giving the stride for reading in the time axis}
\item{deaccumFromFirst}{NULL if no deaccumulation is performed. TRUE or FALSE if deaccumulation is performed from
the first time of the runtime axis or not respectively. If FALSE, an additional runtime is added at the beginning
of each element of the runTimeList to avoid losing the first day when performing deaccumulation.}
\item{doDailyMean}{Logical. Are the forecast time values going to be used for data aggregation?. This argument is passed
to \code{makeSubset} family functions to undertake the pertinent aggregation if TRUE.}
}
\description{
Calculates parameters for forecast time determination along the runtime axis.
}
\author{
J.Bedia, S. Herrera
}
\keyword{internal}

