% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunTimeDomain.R
\name{getRunTimeDomain}
\alias{getRunTimeDomain}
\title{Run time definition according to the time specifications}
\usage{
getRunTimeDomain(dataset, grid, members, season, years)
}
\arguments{
\item{dataset}{Target hindcast dataset}

\item{grid}{A java \dQuote{GeoGrid} containing the target(leading) variable}

\item{members}{Member selection}

\item{season}{Season selection}

\item{years}{Years selected}
}
\value{
A list of parameters:
\itemize{
\item validMonth An integer in the range [1,12] indicating the month to take the initialization
\item years A vector of years selected
\item season Season
\item year.cross An auxiliary integer value thet indicates the position of year-crossing
 within the season vector. NULL if no year-crossing season has been chosen.
\item runDates a POSIXlt vector of initialization dates
\item runTimeRanges a list of of initialization times of the java class \dQuote{ucar.ma2.ranges}.
}
}
\description{
Define index positions of runtimes. 
Its output is passed to the loading functions (the \code{makeSubset\\.*} and \code{derive*} interfaces).
}
\details{
The function calls to specific subroutines for CFS or System4 requests, given their different
runtime configurations. The function also takes care of selecting the appropriate initialization
in the case of year-crossing seasons
}
\author{
J Bedia \email{joaquin.bedia@gmail.com}
}

