% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetDecadalForecast.R
\name{getSeason}
\alias{getSeason}
\title{Get season from a station or field object}
\usage{
getSeason(obj)
}
\arguments{
\item{obj}{Any object extending the station or field classes}
}
\value{
An integer vector with the season
}
\description{
Retrieves the season encompassed by a station or field object
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}

