% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetDecadalForecast.R
\name{getYearsAsINDEX}
\alias{getYearsAsINDEX}
\title{Get years as a factor}
\usage{
getYearsAsINDEX(obj)
}
\arguments{
\item{obj}{Any object extending the station or field classes}
}
\value{
A vector of years of the same length as the time dimension of the object, 
seasonally-adjusted in the case of year-crossing seasons (e.g. DJF). See details.
}
\description{
Extract the year as a factor (e.g. for computing annual statistics)
}
\details{
The function performs a very basic operation, extracting the year element from the 
dates previously converted to POSIXlt. The trick lies in the year-crossing seasons. For instance:
 by convention, winter 2001 encompasses December 2000 and January, February 2001. Therefore, in order to compute
annual statistics for a year-crossing season, it is necessary to modify first the vector of years, 
and assign year 2001 to the preceding December. Similarly, the next December 2001 belongs to winter 2002,
 and so on... The function is useful for computing and/or plotting annual statistics, seasonal climatologies ...
}
\section{Warning}{

The function should no be used to extract the actual years vector
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}

