% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{huss2hurs}
\alias{huss2hurs}
\title{huss2hurs}
\source{
PEcAn Project Repository: 
\url{https://github.com/PecanProject/pecan/blob/master/modules/data.atmosphere/R/metutils.R}
}
\usage{
huss2hurs(huss, tas, ps)
}
\arguments{
\item{huss}{specific humidity, dimensionless (e.g. kg/kg) ratio of water mass / total air mass}

\item{tas}{temperature in degrees C}

\item{ps}{pressure in mb}
}
\value{
relative humidity (in \%), ratio of actual water mixing ratio to saturation mixing ratio
}
\description{
Convert specific humidity to relative humidity
}
\details{
Converting specific humidity into relative humidity from pressure and temperature

\url{http://www.eol.ucar.edu/projects/ceop/dm/documents/refdata_report/eqns.html}
}
\author{
David LeBauer  \email{dlebauer@illinois.edu},
 with a minor modification by J. Bedia to return as \%.
}
\keyword{internal}

