% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{labels_readeR}
\alias{labels_readeR}
\title{Returns a vector of keys to translate with tr.}
\usage{
labels_readeR()
}
\value{
text
}
\description{
Returns a vector of keys to translate with tr.
}
\examples{
labels_readeR()

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
