% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSubset.decadal.R
\name{makeSubset.decadal}
\alias{makeSubset.decadal}
\title{Makes a logical subset of a System4 GeoGrid}
\usage{
makeSubset.decadal(grid, latLon, runTimePars, memberRangeList, foreTimePars,
  verticalPars)
}
\arguments{
\item{grid}{An input java GeoGrid}

\item{latLon}{A list of geolocation parameters, as returned by getLatLonDomainForecast}

\item{runTimePars}{A list of run time definition parameters, as returned by getRunTimeDomain.decadal}

\item{memberRangeList}{A list of ensemble java ranges as returned by getMemberDomain.S4}

\item{foreTimePars}{A list of forecast time definition parameters, as returned by getForecastTimeDomain.S4}

\item{verticalPars}{A list with vertical level definition, as returned by \code{getVerticalLevelPars.ECOMS}. Only the last element (the kava index position) is used}
}
\value{
A list containing a n-dimensional array and the (possibly) modified foreTimePars object with the 
corrected dates in case of time aggregations. Dimensions are labelled by the \dQuote{dimnames} attribute.
}
\description{
Makes a logical subset of a System4 GeoGrid using the parameters specified by the user,
applying the java methods makeSubset and readDataSlice.
}
\details{
Dimensions of length one are dropped and the \dQuote{dimnames} attribute is consequently modified.
In the current version the Z dimension is ignored (and dropped), as it is not planned to include multi-level variables
in the ECOMS-UDG by the moment.
}
\author{
J Bedia, A. Cofino
}
\references{
\url{http://www.unidata.ucar.edu/software/thredds/v4.3/netcdf-java/v4.3/javadocAll/ucar/nc2/dt/grid/GeoGrid.html}
}
\keyword{internal}

