% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSubset.seasonal.R
\name{makeSubset.seasonal}
\alias{makeSubset.seasonal}
\title{Makes a logical subset of a seasonal forecast GeoGrid}
\usage{
makeSubset.seasonal(grid, latLon, runTimePars, memberRangeList, foreTimePars,
  verticalPars)
}
\arguments{
\item{grid}{An input java GeoGrid}

\item{latLon}{A list of geolocation parameters, as returned by getLatLonDomainForecast}

\item{runTimePars}{A list of run time definition parameters, as returned by getRunTimeDomain}

\item{memberRangeList}{A list of ensemble java ranges as returned by getMemberDomain}

\item{foreTimePars}{A list of forecast time definition parameters, as returned by getForecastTimeDomain.S4}

\item{verticalPars}{A list with vertical level definition, as returned by \code{getVerticalLevelPars}.
Only the last element (the java index position) is used}
}
\value{
A list containing a n-dimensional array and the (possibly) modified foreTimePars object with the 
corrected dates in case of time aggregations. Dimensions are labelled by the \dQuote{dimnames} attribute.
}
\description{
Makes a logical subset of a seasonal forecast GeoGrid using the parameters specified by the user,
applying the java methods makeSubset and readDataSlice. Subroutine of loadSeasonalForecast.
}
\details{
Dimensions of length one are dropped and the \dQuote{dimnames} attribute is consequently modified.
In the current version the Z dimension is ignored (and dropped),
 as it is not planned to include multi-level variables.
}
\author{
J Bedia, A.Cofino
}
\references{
\url{http://www.unidata.ucar.edu/software/thredds/v4.3/netcdf-java/v4.3/javadocAll/ucar/nc2/dt/grid/GeoGrid.html}
}
\keyword{internal}

