% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_carga_datos.R
\name{mod_carga_datos_server}
\alias{mod_carga_datos_server}
\title{carga_datos Server Functions}
\usage{
mod_carga_datos_server(id, updateData)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{updateData}{shiny reactive values.}
}
\value{
shiny server
}
\description{
carga_datos Server Functions
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
