% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_carga_datos.R
\name{mod_carga_datos_ui}
\alias{mod_carga_datos_ui}
\title{carga_datos UI Function}
\usage{
mod_carga_datos_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\value{
shiny ui
}
\description{
carga_datos UI Function
}
\examples{
mod_carga_datos_ui("carga_datos_ui_1")

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
