% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDataset.R
\name{openDataset}
\alias{openDataset}
\title{Open a grid dataset}
\usage{
openDataset(dataset)
}
\arguments{
\item{dataset}{The character string defining the dataset location}
}
\value{
A java object of the class GridDataset
}
\description{
Open a local or remote grid dataset
}
\author{
J. Bedia, A Cofino, M. Vega
}
\keyword{internal}

