% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{options.run}
\alias{options.run}
\title{Returns HTML for a Run button.}
\usage{
options.run(runid)
}
\arguments{
\item{runid}{The input slot that will be used to access the value.}
}
\value{
HTML
}
\description{
Returns HTML for a Run button.
}
\examples{
options.run("id")

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
