% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readeR.R
\docType{package}
\name{readeR}
\alias{readeR}
\title{Read Data System}
\description{
Provides a framework for a database through a 'shiny' graphical interface.
}
\details{
\tabular{ll}{
Package: \tab readeR\cr
Type: \tab Package\cr
Version: \tab 1.0.4\cr
Date: \tab 2022-09-30\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Maintainer: Oldemar Rodriguez Rojas <oldemar.rodriguez@ucr.ac.cr>\cr
\itemize{
  \item Oldemar Rodriguez Rojas <oldemar.rodriguez@ucr.ac.cr>
  \item Diego Jiménez Alvarado
  \item Joseline Quirós Mendez
}
}
\keyword{package}
