% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadStationData.R
\name{string2date}
\alias{string2date}
\title{POSIXlt conversion from character}
\usage{
string2date(timeString, tz = tz)
}
\arguments{
\item{timeString}{Date vector as stored in VALUE files, previously coerced to character}

\item{tz}{Time zone. See \code{\link{loadStationData}}}
}
\value{
A POSIXlt vector of the same length of the input
}
\description{
Converts the date codes of the Value format to \code{"POSIXlt"}
}
\details{
Currently the VALUE format is intended for daily data of the form YYYMMDD. However,
 the function also considers the possibility of subdaily data if hourly data are introduced in
 the form YYYYMMDDHH, eading to a string of 10 characters.
}
\note{
The function is currently exported to be internally used by the VALUE package to load predictions.
}
\author{
juaco
}
\keyword{internal}
