% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subsetDecadalForecast.R
\name{subsetDecadalForecast}
\alias{subsetDecadalForecast}
\title{Subset decadal forecast by initialization years}
\usage{
subsetDecadalForecast(data, ly = 1)
}
\arguments{
\item{data}{field obtained from the output of function loadDecadalForecast.}

\item{ly}{integer indicating the lead year of initialization}
}
\description{
Subset decadal forecast by initialization years
}
\examples{
\dontrun{
latLim <- c(-20,10)
lonLim <-  c(-60,-30)
season <- 3:5
period <- 1981:1991
loginUDG(username = "myuser", password = "mypassword") #type help(loginUDG)
tasDECA <- loadDecadalForecast(dataset = "http://www.meteo.unican.es/tds5/dodsC/specs/gfdl_specs_decadal.ncml",
   latLim = latLim, lonLim = lonLim,
   var = "tas", dictionary = F, years = 1981:1982, season = season)
# data corresponding to the initialitation of the year before
subsetDecadalForecast(tasDECA, ly = 1)
}
}
\author{
J. Bedia, S. Herrera, M. Iturbide, J.M. Gutierrez
}

