% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{tas_decadalForecast}
\alias{tas_decadalForecast}
\title{Field of decadal foecast data from the SPECS_GFDL_decadal dataset containing monthly temperature for the Ibrian Peninsula.}
\format{Field data}
\source{
Subset of the SPECS_GFDL_decadal dataset. This data is accessible 
for loading in the \strong{The User Data Gateway (UDG)}, which is the one stop shop for 
climate data access maintained by the Santander MetGroup. 
The UDG builds on the THREDDS Access Portal (UDG-TAP) which is the entry point for 
authentication and data access (more info in \strong{https://meteo.unican.es/trac/wiki/udg}).
}
\description{
latLim <- c(35,45)
lonLim <-  c(-10,2)
season <- 3:5
period <- 1981:1982
var= "tas".
}
\keyword{SPECS}
\keyword{decadal}
\keyword{forecast}
\keyword{temperature}

