% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{translation.readeR}
\alias{translation.readeR}
\title{Returns a list of sentences with their translation in different languages.}
\usage{
translation.readeR()
}
\value{
list
}
\description{
Returns a list of sentences with their translation in different languages.
}
\examples{
translation.readeR()

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
