% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{UDG.datasets}
\alias{UDG.datasets}
\title{Show UDG datasets}
\usage{
UDG.datasets()
}
\value{
The datasets table, in the form of a data.frame
}
\description{
Access the installed user's vocabulary
}
\note{
The function assumes that the user has read permission to the package installation directory
}
\examples{
# Default built-in datasets
(sets <- UDG.datasets())
sets[grep("^ta", sets$name), ]
}
\author{
M Iturbide
}
