% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{UDG.vocabulary}
\alias{UDG.vocabulary}
\title{Show UDG vocabulary}
\usage{
UDG.vocabulary()
}
\value{
The vocabulary table, in the form of a data.frame
}
\description{
Access the installed user's vocabulary
}
\note{
The function assumes that the user has read permission to the package installation directory
}
\examples{
# Default built-in vocabulary
(voc <- UDG.vocabulary())
voc[grep("^ta", voc$identifier), ]
}
\author{
J Bedia
}
\seealso{
UDG.vocabulary.update, for the inclusion of new standard variables defined by the user
}

