% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeStationData.R
\name{writeStationData}
\alias{writeStationData}
\title{Write to Value format}
\usage{
writeStationData(obj, path, na.code = NA)
}
\arguments{
\item{obj}{Object as returned from \code{\link[loadeR]{loadStationData}}.}

\item{path}{A string. Specify the path where to save the .txt.}

\item{na.code}{Possible values are NA or NaN. Default is NA. Indicates how missing values should be specified.}
}
\description{
Write to Value format
}
\author{
J. Bano
}
