% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustDates.forecast.R
\name{adjustDates.forecast}
\alias{adjustDates.forecast}
\title{Date adjustment}
\usage{
adjustDates.forecast(foreTimePars)
}
\arguments{
\item{foreTimePars}{A list of elements as returned by \code{getRunTimeDomain.decadal}}
}
\description{
Adjust dates in forecast data
}
\author{
J. Bedia, S. Herrera
}
\keyword{internal}

