% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLatLonDomain.R
\name{adjustRCMgrid}
\alias{adjustRCMgrid}
\title{RCM grid adjustment}
\usage{
adjustRCMgrid(gds, latLon, lonLim, latLim)
}
\arguments{
\item{gds}{a Java GridDataSet object}

\item{latLon}{latLon definition (see \code{\link{getLatLonDomain}})}

\item{lonLim}{lonLim}

\item{latLim}{latLim}
}
\value{
a new latLon definition
}
\description{
Performs operations to adequately handle 2D XY axis (typically from RCMs)
}
\author{
S. Herrera, M. Iturbide, J. Bedia
}
\keyword{internal}

