% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaryTransformForecast.R
\name{dictionaryTransformForecast}
\alias{dictionaryTransformForecast}
\title{Performs variable transformation}
\usage{
dictionaryTransformForecast(dic, mdArray)
}
\arguments{
\item{dic}{Dictionary line for the variable, as returned by \code{dictionaryLookup.ECOMS}}

\item{mdArray}{A n-dimensional array, as returned by \code{makeSubset}}
}
\value{
The transformed n dimensional array of data. See details.
}
\description{
Performs variable transformation according to dictionary specifications.
Sub-routine of the \code{load*Forecast} methods.
}
\author{
J. Bedia
}
\keyword{internal}

