% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{options.run}
\alias{options.run}
\title{Creates a button to use in a options menu.}
\usage{
options.run(runid, label = "opts")
}
\arguments{
\item{runid}{The input slot that will be used to access the value.}

\item{label}{A label on the top of the menu.}
}
\value{
An HTML element.
}
\description{
Creates a button to use in a options menu.
}
\examples{
options.run("id")

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
