% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readeR.R
\docType{package}
\name{readeR}
\alias{readeR}
\title{Read Data System}
\description{
System to read data and show it.
}
\details{
\tabular{ll}{
Package: \tab readeR\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 2022-01-20\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Oldemar Rodriguez Rojas \cr
Maintainer: Oldemar Rodriguez Rojas <oldemar.rodriguez@ucr.ac.cr>
}
\keyword{package}
