% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{UDG.datasets}
\alias{UDG.datasets}
\title{Show UDG datasets}
\usage{
UDG.datasets(pattern = "")
}
\arguments{
\item{pattern}{Optional. Pattern in the dataset name as passed to function \code{\link{grep}} (case-insensitive).}
}
\value{
The datasets table, in the form of a \code{data.frame}
}
\description{
Access the installed user's vocabulary
}
\note{
The function assumes that the user has read permission to the package installation directory
}
\examples{
# Default built-in datasets
str(UDG.datasets())
(sets <- UDG.datasets())
sets[grep("CORDEX-EUR44.*historical", sets$name), ]
# Using argument pattern
UDG.datasets(pattern = "CORDEX-EUR44.*historical")$name
}
\author{
M Iturbide
}
